requirements_xkx.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.3.1
altair==5.1.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backoff==1.11.1
basicsr==1.4.2
bcrypt==4.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.8.2
build==1.0.3
cachetools==5.5.0
certifi==2024.6.2
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.16
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
cos-python-sdk-v5==1.9.31
coverage==7.4.1
crcmod==1.7
cycler==0.11.0
Cython==3.0.2
dataclasses-json==0.6.6
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docutils==0.20.1
docx2txt==0.8
durationpy==0.9
easydict==1.10
email_validator==2.2.0
exceptiongroup==1.1.3
executing==2.1.0
facexlib==0.3.0
fastapi==0.110.3
fastapi-cli==0.0.5
feedparser==6.0.11
ffmpy==0.3.1
filelock==3.14.0
filterpy==1.4.5
flake8==7.0.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.6.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.43
gkeepapi==0.16.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
gpsoauth==1.1.1
gradio==3.47.1
gradio_client==0.6.0
grandalf==0.8
greenlet==3.0.3
gritql==0.1.5
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.3
humanfriendly==10.0
hyperframe==6.0.1
identify==2.5.33
idna==3.7
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
inquirer==3.1.3
insightface==0.7.3
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.13.1
keyring==24.3.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.3
langchain-chroma==0.1.4
langchain-cli==0.0.31
langchain-community==0.3.2
langchain-core==0.3.11
langchain-experimental==0.3.1
langchain-openai==0.2.2
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langserve==0.3.0
langsmith==0.1.135
lazy_loader==0.3
libclang==16.0.6
libcst==1.4.0
llvmlite==0.40.1
lmdb==1.4.1
lxml==5.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nh3==0.2.15
nodeenv==1.8.0
numba==0.57.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.16.0
openai==1.51.2
openai-token-counter==1.0.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.76
opennsfw2==0.10.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.9.6
platformdirs==4.2.2
playwright==1.41.1
pluggy==1.4.0
portalocker==2.8.2
posthog==3.7.0
pre-commit==3.6.0
prettytable==3.9.0
prompt_toolkit==3.0.47
protobuf==5.27.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic-settings==2.5.2
pydantic_core==2.18.4
pydub==0.25.1
pyee==11.0.1
pyflakes==3.2.0
Pygments==2.18.0
pylint-venv==3.0.3
pyparsing==3.1.1
pypdf==4.2.0
PyPika==0.48.9
pyproject-api==1.6.1
pyproject-toml==0.0.10
pyproject_hooks==1.0.0
PySocks==1.7.1
pytest==8.0.0
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.9
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.10.1
qudida==0.0.4
readchar==4.0.5
readme-renderer==42.0
realesrgan==0.3.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tb-nightly==2.14.0a20230808
tcvdb-text==1.0.0
tcvectordb==1.4.0
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.18
tiktoken==0.7.0
tokencost==0.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.5
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.4.1
tox==4.12.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
twine==4.0.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.23.2
uvloop==0.20.0
virtualenv==20.25.0
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0.1
Werkzeug==3.0.3
wrapt==1.15.0
xmltodict==0.13.0
yapf==0.40.1
yarl==1.9.4
zipp==3.17.0
